Mistake
I've made a mistake. Restored the code to commit 'End of video 9'
